Update Week 1&2
Initially my understanding for  ord() is that is an abbreviation of ordinal: it defines numerically the position of something in a series, making it first, second, third and so on, however on learning and researching more i have come to realise that the ord() function is a built in function within Python. The function returns an integer value for a single Unicode character.